Project_material
# nce-project-material
this project will handle project material for nce and help them to download project material in doc
the issue is that the footer dont stick to the bottom i try if their is no content on the page the footer move to the top. so i want it to stick to the bottom on every page. more 
so the sidebar to place i find it hard to feed it with data from list_of_department.